Mod start function will never return undefined for a minimal mod configuration
It could have happened before that using the minimal configuration, a module is registered correctly but would then fail to start, throwing "Error: Module should return a valid object, to be used by other modules". This is now fixed, by ensuring that when an explicit obj is not available, the start function returns the mod object instead. This is now tested explicitly in the test suite.